Eliminate stutter while pausing and using OSS (or libao, which uses OSS on *BSD)
The problem is that MusiXCurses uses SIGALRM to be awoken every second so it can
do UI updates. However, unsuspended POSIX threads will _also_ be awoken by this
signal, which causes them to break free of the condition variable wait loop.

The result is that they decode a piece of information, but immediately suspend
themselves thereafter, pausing audio output. Between the awaken and suspend
events, they manage to squeeze a tiny bit of audio out, which causes the stutter
heard.