Fix a deadlock that would occur if the player were to terminate while the buffer was filled.
As the filler thread would only check for termination after waiting for the 'data has been read'
condition variable, it'd be stuck forever since the buffer will never be read. Likewise, the
destructor would wait for the thread to exit, which it never does.

The issue is resolved by checking for available buffer space *or* termination, and doing an
explicit signal of the condition variable to ensure the filler thread breaks free.