If a playlist entry has completed, remove it from the playlist.
This has the result of keeping a concise playlist in the player, with the downside of utterly discarding any history. The main issue is that it seems impossible to fit the playlist on the screen, annotate the tracks and have room left for options; so this seems to be the only compromise I can think of.